requirements.txt
blinker==1.6.2
click==8.1.5
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
zipp==3.16.2
Flask==2.3.2
flask-migrate==3.1.0
sqlalchemy==1.4.25
flask-sqlalchemy==2.5.1
sqlalchemy-utils==0.37.8
python-dateutil==2.8.2
